Work on gate statements, add PacketBufferBC for writing 8 booleans into a byte automatically.
A bit of work on fluid pipes.

+++ common/buildcraft/transport/plug/PluggablePulsar.java
@@ -74,7 +74,7 @@ public class PluggablePulsar extends PipePluggable {
     public NBTTagCompound writeToNbt() {
         NBTTagCompound nbt = super.writeToNbt();
         nbt.setBoolean("isPulsing", isPulsing);
-        nbt.setInteger("gateEnabledTicks", (byte) gateEnabledTicks);
+        nbt.setInteger("gateEnabledTicks", gateEnabledTicks);
         nbt.setInteger("gateSinglePulses", gateSinglePulses);
         return nbt;
     }
